Enable local object lookup (in DFC)
Problem: Currently, the HEAD object API does not tell you if an object from a cloud bucket is cached in DFC (this is not only a useful API to have, but required for the multi-tier DFC strategy).

Solution: A new query string param is being introduced to enabled this. Unfortunately, a bit of refactoring is required here to decouple the function lookupLocally from the type (i.e. local or not) of bucket being examined.