Do not use EC for bucket that do not have properties.
It is usually cloud buckets that should not be EC'ed.

Allow to use lower values for slice count (now minimal
number of data and parity slices is 1

Tests auto-detects the number of data and parity slices
depending on the number of target. So, both value are
between 1 and 2 (the smallest number of target to run
tests is 4 now - was 6 before)