Downloader: introduce DownloaderJob interface
DownloaderJob interface is an API which downloader needs to start
downloading. This will allow dispatcher to calculate/fetch object names
to download in batches, which will allow to prevent overloading
downloader's joggers in the future.

This change changes a bit Download() operation. Previously this
operation was waiting until all objects' loms were successfully created.
As this could have taken a lot of time for large datasets, this part is removed
and Download() operation terminates almost immediately. If any errors
occured, they will be visible in status request