sdk/python: Implement BatchLoader + getbatch API for batch data loading
`BatchLoader` allows for the download of data stored across multiple objects in an AIStore cluster. This includes across multiple buckets/providers as well. Batch requests can be served as a stream containing an archive or as a multipart response (part 1 being metadata, part 2 being the archive).

* [x] Investigate how this can be used in the DataLoader
* [x] Use AIS client and retry support
* [x] Implement reader support for other mime types (atleast json.gz for now, zip and jsonlz4 later)
* [x] ~~Consider chunking reads for large files~~ (Future patch instead)
* [x] Implement unit tests
* [x] Implement single-node integration tests

Depends on the multipart decoder patch.

Signed-off-by: Soham Manoli <msoham123@gmail.com>